amelioration again
